Nix suggestion that pipelined increases clock speed
A reader pointed out that this is a bit misleading. While pipelining a
CPU does increase its throughput (since each stage can execute in a
shorter clock cycle than if an instruction had to make it through the
whole CPU in one go), pipelining can only slow the time taken for an
instruction to get from start to end.